refactor: searchParams 비동기 처리 및 클라이언트 컴포넌트 분리 (#114)
